Update bash
Update bash to be able to make file without moving files

+++ Code/test.sh
@@ -7,16 +7,18 @@ clear
 make clean
 make
 
-# Test and play
-echo "------------------------------------------------------------------"
-echo "./PDGWB"
-./PDGwB
-# Update the files in the Local Git Files
-# Remove Old Files
-rm ../sf/PDGwB/*
-# Replace them ???
-files=`ls`
-for eachFile in $files
-do
- cp $eachFile ../sf/PDGwB/Code
-done
+if(($1 ==0));then
+ # Test and play
+ echo "------------------------------------------------------------------"
+ echo "./PDGWB"
+ ./PDGwB
+ # Update the files in the Local Git Files
+ # Remove Old Files
+ rm ../sf/PDGwB/*
+ # Replace them ???
+ files=`ls`
+ for eachFile in $files
+ do
+  cp $eachFile ../sf/PDGwB/Code
+ done
+fi
